ai-engine/requirements.txt
tensorflow==2.12.0
tensorflow-probability==0.20.1
numpy==1.24.3
scikit-learn==1.2.2
kafka-python==2.0.2
pandas==2.0.2
matplotlib==3.7.1
scipy==1.10.1
pyopenssl==23.2.0
hyperopt==0.2.7
mlflow==2.3.1